console middleware update
